requirements.txt
Package                   Version
------------------------- -----------
absl-py                   1.4.0
aiofiles                  23.1.0
aiohttp                   3.8.4
aiosignal                 1.3.1
altair                    4.2.2
altgraph                  0.17.3
anyio                     3.6.2
astor                     0.8.1
async-timeout             4.0.2
asynctest                 0.13.0
attrdict                  2.0.1
attrs                     22.2.0
Babel                     2.12.1
bce-python-sdk            0.8.79
Brotli                    1.0.9
cachetools                4.2.4
certifi                   2022.12.7
cffi                      1.15.1
charset-normalizer        3.1.0
click                     8.1.3
colorama                  0.4.6
coloredlogs               15.0.1
cssselect                 1.2.0
cssutils                  2.6.0
cycler                    0.11.0
Cython                    0.29.33
ddddocr                   1.4.7
decorator                 5.1.1
entrypoints               0.4
et-xmlfile                1.1.0
fastapi                   0.94.1
ffmpy                     0.3.0
filelock                  3.9.1
Flask                     2.2.3
flask-babel               3.0.1
flatbuffers               2.0
fonttools                 4.38.0
frozenlist                1.3.3
fsspec                    2023.1.0
future                    0.18.3
gast                      0.2.2
gevent                    22.10.2
geventhttpclient          2.0.2
glob3                     0.0.1
google-auth               1.35.0
google-auth-oauthlib      0.4.6
google-pasta              0.2.0
gradio                    3.21.0
greenlet                  2.0.2
grpcio                    1.52.0
h11                       0.14.0
h5py                      2.10.0
httpcore                  0.16.3
httpx                     0.23.3
huggingface-hub           0.13.2
humanfriendly             10.0
idna                      3.4
imageio                   2.26.0
imgaug                    0.4.0
importlib-metadata        6.0.0
importlib-resources       5.12.0
itsdangerous              2.1.2
Jinja2                    3.1.2
jsonschema                4.17.3
Keras-Applications        1.0.8
Keras-Preprocessing       1.1.2
keyboard                  0.13.5
kiwisolver                1.4.4
lanms-neo                 1.0.2
linkify-it-py             2.0.0
lmdb                      1.4.0
lxml                      4.9.2
Markdown                  3.4.1
markdown-it-py            2.2.0
MarkupSafe                2.1.2
matplotlib                3.5.3
mdit-py-plugins           0.3.3
mdurl                     0.1.1
MouseInfo                 0.1.3
mpmath                    1.2.1
multidict                 6.0.4
networkx                  2.6.3
numpy                     1.21.6
oauthlib                  3.2.2
onnx                      1.13.1
onnxruntime               1.12.1
opencv-contrib-python     4.5.5.62
opencv-python             4.5.5.62
openpyxl                  3.1.2
opt-einsum                3.3.0
orjson                    3.8.7
packaging                 23.0
paddle-bfloat             0.1.7
paddlepaddle              2.4.2
pandas                    1.3.5
pefile                    2023.2.7
Pillow                    9.4.0
pip                       23.0.1
pkgutil_resolve_name      1.3.10
Polygon3                  3.0.9.1
premailer                 3.10.0
protobuf                  3.20.3
psutil                    5.9.4
pyasn1                    0.4.8
pyasn1-modules            0.2.8
PyAutoGUI                 0.9.53
pyclipper                 1.3.0.post4
pycparser                 2.21
pycryptodome              3.17
pydantic                  1.10.6
PyDirectInput             1.0.4
pydub                     0.25.1
PyGetWindow               0.0.9
pyinstaller               5.9.0
pyinstaller-hooks-contrib 2023.1
PyMsgBox                  1.0.9
PyMuPDF                   1.20.2
pyparsing                 3.0.9
pyperclip                 1.8.2
PyQt5                     5.15.9
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.11.1
pyreadline                2.1
PyRect                    0.2.0
pyrsistent                0.19.3
PyScreeze                 0.1.28
python-dateutil           2.8.2
python-multipart          0.0.6
python-rapidjson          1.9
pytweening                1.0.4
pytz                      2022.7.1
PyWavelets                1.4.0
pywin32                   305
pywin32-ctypes            0.2.0
PyYAML                    6.0
rapidfuzz                 2.13.7
rarfile                   4.0
requests                  2.28.2
requests-oauthlib         1.3.1
rfc3986                   1.5.0
rsa                       4.9
scikit-image              0.19.3
scipy                     1.4.1
setuptools                67.6.0
shapely                   2.0.1
six                       1.16.0
sniffio                   1.3.0
starlette                 0.26.1
sympy                     1.10.1
tensorboard               2.1.1
tensorflow-gpu            2.1.0
tensorflow-gpu-estimator  2.1.0
termcolor                 2.2.0
tifffile                  2021.11.2
toolz                     0.12.0
tqdm                      4.65.0
tritonclient              2.31.0
typing_extensions         4.5.0
uc-micro-py               1.0.1
urllib3                   1.26.15
uvicorn                   0.21.0
visualdl                  2.5.1
websockets                10.4
Werkzeug                  2.2.3
wheel                     0.40.0
wrapt                     1.15.0
x2paddle                  1.4.0
yarl                      1.8.2
zipp                      3.15.0
zope.event                4.6
zope.interface            5.5.2
